rename CommonQIR to QIRCommon
